Fix : 데이터 형식 변환

+++ src/main/java/com/sdy/bbbb/data/DataService.java
@@ -133,7 +133,7 @@ public class DataService {
             for(SpotCalculated spot1 : spotCalculateds) {
                 if (guBaseInfo.getArea_nm().equals(spot1.getArea_Nm())) {
 
-                    lastPopByHour.put("L" + spot1.getThat_Hour(), Long.parseLong(spot1.getPopulation_By_Hour()));
+                    lastPopByHour.put("L" + spot1.getThat_Hour(), (long)Double.parseDouble(spot1.getPopulation_By_Hour()));
 //                    mapList.add(popByHour);
                 }
             }
@@ -142,7 +142,7 @@ public class DataService {
             for(SpotCalculated spot2 : todaySpotCalculatedList) {
                 if (guBaseInfo.getArea_nm().equals(spot2.getArea_Nm())) {
 
-                    todayPopByHour.put("T" + spot2.getThat_Hour(), Long.parseLong(spot2.getPopulation_By_Hour()));
+                    todayPopByHour.put("T" + spot2.getThat_Hour(), (long)Double.parseDouble(spot2.getPopulation_By_Hour()));
 //                    mapTodayList.add(todayPopByHour);
                 }
             }
